+++ Ya/Algebra/Definition.hs
@@ -148,14 +148,16 @@ type family Neutral p where
 	Neutral (/\) = Unit
 	Neutral (\/) = Void
 
-type Day = U_UU_UUU_UUUU_T_TT_I_II_III (/\)
+type Day = U_V_UU_UUU_UUUU_T_TT_I_II_III (/\)
 
 class
-	( Mapping v from into (Day (v from) p pp f f i ii) f
+	( Mapping v from into (Day v from p pp f f i ii) f
 	, Mapping v from into (v from (Neutral p)) f
+	, Dumb (x v from into f)
 	) => Monoidal x v p pp from into i ii f
 
 deriving instance
-	( Mapping v from into (Day (v from) p pp f f i ii) f
+	( Mapping v from into (Day v from p pp f f i ii) f
 	, Mapping v from into (v from (Neutral p)) f
+	, Dumb (x v from into f)
 	) => Monoidal Functor v p pp from into i ii f